pyproject.toml
[project]
name = "offramp"
version = "0.1.0"
description = "Salesforce Off-Ramp: extract, translate, validate, and migrate Salesforce automation onto an external runtime."
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "Proprietary" }
authors = [{ name = "The Attic AI, Inc." }]

# Phase 0 keeps runtime deps minimal — they grow per-phase.
# See docs/architecture.md §4 for the full target dependency list.
dependencies = [
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "structlog>=24.1",
    "anyio>=4.4",
    "httpx>=0.27",
]

[project.scripts]
offramp = "offramp.cli.__main__:main"

[project.optional-dependencies]
# Loaded lazily as later phases come online.
salesforce = [
    "simple-salesforce>=1.12.9",
]
graph = [
    "falkordb>=1.0",
    "networkx>=3.3",
]
llm = [
    "anthropic>=0.40",
]
runtime = [
    "temporalio>=1.16",
]
report = [
    "jinja2>=3.1",
]
pubsub = [
    "grpcio>=1.66",
    "grpcio-tools>=1.66",
    "protobuf>=5.27",
    "fastavro>=1.9",
]
shadow = [
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "pre-commit>=3.7",
    "respx>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/offramp"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]
# Generated protobuf stubs and infra dirs are excluded from lint/format.
extend-exclude = [
    "infra",
    "docs",
    "src/offramp/validate/shadow/pubsub_proto/pubsub_api_pb2.py",
    "src/offramp/validate/shadow/pubsub_proto/pubsub_api_pb2_grpc.py",
]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "B",             # bugbear
    "UP",            # pyupgrade
    "SIM",           # simplify
    "RUF",           # ruff-specific
    "ASYNC",         # async best practices
]
ignore = [
    "E501",          # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]                     # allow assert False in tests
"src/offramp/cli/**" = ["B008"]           # CLI default-arg patterns

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_decorators = true
files = ["src", "tests"]
exclude = [
    "infra/.*",
    "src/offramp/validate/shadow/pubsub_proto/pubsub_api_pb2.py",
    "src/offramp/validate/shadow/pubsub_proto/pubsub_api_pb2_grpc.py",
]

# 3rd-party libs without bundled stubs — relax until each phase introduces them.
[[tool.mypy.overrides]]
module = [
    "simple_salesforce.*",
    "falkordb.*",
    "temporalio.*",
    "structlog.*",
    "networkx.*",
    "asyncpg.*",
    "grpc.*",
    "fastavro.*",
    "offramp.validate.shadow.pubsub_proto.*",
]
ignore_missing_imports = true

# The pubsub gRPC client uses dynamic attributes from the generated stubs.
# Runtime works; mypy can't see through generated protobuf modules.
[[tool.mypy.overrides]]
module = ["offramp.validate.shadow.pubsub"]
disable_error_code = [
    "attr-defined",
    "name-defined",
    "type-arg",
    "no-untyped-call",
]

# Tests use pytest fixtures + lambdas heavily — relax annotation-required
# rules for tests/* while keeping strict typing for src/*. Real bugs still
# get caught: argument-type mismatches, attribute errors, etc.
[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: requires external services (scratch org, Postgres, etc.)",
    "smoke: end-to-end smoke tests",
    "ooe: OoE runtime test suite",
    "load: performance benchmarks",
]

[tool.coverage.run]
source = ["src/offramp"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80
